refactor: (#801) exception 통일
