--- dms-core/src/main/kotlin/team/aliens/dms/domain/vote/exception/NotValidPeriodException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.vote.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.vote.exception.error.VotingTopicErrorCode
-
-object NotValidPeriodException : DmsException(
-    VotingTopicErrorCode.NOT_VALID_PERIOD
-)
