--- dms-core/src/main/kotlin/team/aliens/dms/domain/vote/exception/NotVotingPeriodException.kt
@@ -1,8 +1,0 @@
-package team.aliens.dms.domain.vote.exception
-
-import team.aliens.dms.common.error.DmsException
-import team.aliens.dms.domain.vote.exception.error.VotingTopicErrorCode
-
-object NotVotingPeriodException : DmsException(
-    VotingTopicErrorCode.NOT_VOTING_PERIOD
-)
