+++ dms-core/src/main/kotlin/team/aliens/dms/domain/vote/exception/VotingTopicErrorCode.kt
@@ -1,4 +1,4 @@
-package team.aliens.dms.domain.vote.exception.error
+package team.aliens.dms.domain.vote.exception
 
 import team.aliens.dms.common.error.ErrorProperty
 import team.aliens.dms.common.error.ErrorStatus
@@ -10,7 +10,7 @@ enum class VotingTopicErrorCode(
 ) : ErrorProperty {
 
     NOT_VALID_PERIOD(ErrorStatus.BAD_REQUEST, "Not Valid Period", 1),
-    NOT_FOUND_VOTING_TOPIC(ErrorStatus.BAD_REQUEST, "Not Found VotingTopic", 2),
+    VOTING_TOPIC_NOT_FOUND(ErrorStatus.BAD_REQUEST, "VotingTopic Not Found", 2),
     NOT_VOTING_PERIOD(ErrorStatus.BAD_REQUEST, "Not Voting Period", 3),
     ;
